requirements.txt
docx-mailmerge==0.3.0
lxml==3.7.3
numpy==1.12.0
pandas==0.19.2
pyodbc==4.0.14
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
